+++ hw2/Picture.h
@@ -13,7 +13,7 @@ public:
 
     bool read(std::ifstream &in);
     bool write(std::ofstream &out);
-    void drawLine(long double x1, long double y1, long double x2, long double y2, long double thickness, int brightness, GammaCorrection gammaCorrection);
+    void drawLine(double x1, double y1, double x2, double y2, double thickness, int brightness, GammaCorrection gammaCorrection);
 
 private:
     std::vector<uchar> data;
